refactor(src/lib/util/resize.ts): move resize back into the lab, not ready for lib yet
