+++ src/app/lab/04-vectors/main.ts
@@ -1,7 +1,7 @@
 import { el } from '../../../lib/dom';
 import loop from '../../../lib/game/loop';
 import { random, sin, cos, ππ, round } from '../../../lib/math';
-import resize from '../../../lib/util/resize';
+import resize from './resize';
 import { vec2 } from '../../../lib/vec/factories';
 import { sub, mul, add, div } from '../../../lib/vec/math';
 import { Vec2 } from '../../../lib/vec/types';
